dom/base/test/test_bug787778.html

<!DOCTYPE html>
<!--
    https://bugzilla.mozilla.org/show_bug.cgi?id=787778
  -->
<html>
  <head>
    <meta charset="utf-8">
    <title>Test for Bug 787778</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="text/javascript">
      SimpleTest.waitForExplicitFinish();
      window.onload = function() {
        window.document.x.src='a';
        SimpleTest.executeSoon(function () {
          ok(true, "Didn't crash!");
          SimpleTest.finish();
        });
      }
    </script>
  </head>
  <body>
    <embed name="x" src="./file_bug787778.sjs">
  </body>
</html>
